create more dynamic home screen to be used
